Created myRooms page

+++ server.js
@@ -131,40 +131,44 @@ app.get("/logout", (req, res) => {
 });
 
 app.get("/myRooms", async (req, res) => {
-  const userId = req.session.user_id;
-
-  // Query to get all rooms for the current user
-  const query = `
-  SELECT room.room_id, lmd.last_message_date, rm.unread_message_count
-  FROM room_user
-  JOIN room ON room_user.room_id = room.room_id
-  JOIN (
-    SELECT room_user.room_id, MAX(message.sent_datetime) as last_message_date
+  if(!req.session.user_id){
+    res.status(401).send("Unauthorized");
+    return;
+  }else{
+    const userId = req.session.user_id;
+    // Query to get all rooms for the current user
+    const query = `
+    SELECT room.room_id, lmd.last_message_date, rm.unread_message_count
     FROM room_user
-    LEFT JOIN message ON room_user.room_user_id = message.room_user_id
-    GROUP BY room_user.room_id
-  ) lmd ON lmd.room_id = room.room_id
-  JOIN (
-    SELECT room_user.room_id,COUNT(*) as unread_message_count
-    FROM message
-    JOIN room_user ON message.room_user_id = room_user.room_user_id
-    WHERE room_user.last_read_msg_id IS NULL OR message.message_id > room_user.last_read_msg_id
-    GROUP BY room_user.room_id
-  ) rm ON room_user.room_id = rm.room_id
-  WHERE room_user.user_id = ?
-  GROUP BY room.room_id
-  ORDER BY last_message_date DESC;
-    `;
-
+    JOIN room ON room_user.room_id = room.room_id
+    JOIN (
+      SELECT room_user.room_id, MAX(message.sent_datetime) as last_message_date
+      FROM room_user
+      LEFT JOIN message ON room_user.room_user_id = message.room_user_id
+      GROUP BY room_user.room_id
+    ) lmd ON lmd.room_id = room.room_id
+    JOIN (
+      SELECT room_user.room_id,COUNT(*) as unread_message_count
+      FROM message
+      JOIN room_user ON message.room_user_id = room_user.room_user_id
+      WHERE room_user.last_read_msg_id IS NULL OR message.message_id > room_user.last_read_msg_id
+      GROUP BY room_user.room_id
+    ) rm ON room_user.room_id = rm.room_id
+    WHERE room_user.user_id = ?
+    GROUP BY room.room_id
+    ORDER BY last_message_date DESC;
+      `;
+  
     mysqlConnection.query(query, [userId], (error, results, fields) => {
       if (error) {
-        console.error('Error executing query:', error);
+        console.error("Error executing query:", error);
         return;
       }
-      console.log('Query Results:', results);
+      console.log("Query Results:", results);
       res.send(results);
-    
     });
+  }
+
 });
 
 app.get("*", (req, res) => {
